Added background and ability to shoot
As well as what is mentioned in the title I have changed the formatting of the code a bit to fit better with the GDScript style guide. An obstacle I have realized when adding the background is that it is a bit too distracting, so I will need to find a way to make the stuff happening in the level actually pop out more.